uses the `set` operations in `eval_let` to make it observable
Originally, we decided to use `Env.set` so that let bindings
wouldn't be treated as assignments, this makes them transparent
and reduces the overhead of data flow tracking and analysis (since
let-bindings are not really creating any data points, that we have
to track). However, it immediatelly broke some downstream code, so
as a first approximation we decided to make let-bindings observable.