refines the variable theory
We are now distinguishing between three classes of variables:
- physical registers;
- temporary mutable storages (denoted with `#NNN`)
- immutable variables (denoted with `$NNN`), aka let-bound

The immutable variables could be used for let-binding and other forms
of abstraction. They could be obtained only through the
`Var.scoped` function, that has type

     'a sort -> ('a var -> 'b value t) -> 'b value t

which ensures that the given variable will never be modified. The
names of scoped variables are guaranteed to be unique only in the
scope of its definition (in other words, the unqiue counter will
be relesed once the scope is left).

The variable representation now provides an abstract `ident` type,
which could be used to preserve variable kinds during parsing and
unparsing.

Breaking Changes:

- `Var.Generator.fresh` is removed, in favor of two alternatives:
  1. `Var.fresh` which creates an unscoped fresh mutable variable;
  2. `Var.scoped` which creates a scoped immutable variable

- `Var.create` now takes `ident` instead of a string value, another
  function called `Var.define` is provided to define physical
  registers, however the name must be valid.