initiates the migration of BAP Lisp from Primus
This is an investigative branch that will, if succeed, bring the new
semantics representation to BAP.

The idea is simple. First of all, we need a concrete representation of
the semantics, which we can print, store, exchange etc. The
representation must preserve the abstractions provided by the lifters,
be extensible, support hash-consing, etc. In the initial prototype,
this representation was called CIS and had basically the same AST as
Primus Lisp (modulo hash-consing and some other pecularities).

We also need to write a concrete interpreter for the new
representation, which will be able to directly interpret on the
highest abstraction level. Of course, we want to use Primus for
that. And since Primus already contains the (Primus) Lisp interpreter
we can easily reuse Primus Lisp as our main representation.

So it looks like that the majority of the work is already done,
however there still is some impendance mismatch between Primus Lisp
and an ideal candidate for the BAP new semantics language.

First of all, Primus Lisp was assuming BIL as the main bearer of the
semantics, so it relies on Bap.Std.Var and Bap.Std.Type. So we need
to switch to the Core Theory's Var and Sort types.

Second, Primus Lisp is gradually typed, while we need strict typing
for defining the semantics. This will require more type annotations,
as well as declarations of primitives. Which is good, as it will make
the language stronger and easier to use. However, it will involve
some work.

Third, Primus Lisp's program model is interleaved with the
runtime (dynamic) representation of primitives, e.g., the `primitive`
type denotes Primus Machine computation, while the representation
should be completely static. We still need a primitive, since we're
not going to represent bitvector addition concretely, however it is
a (language) design question where to draw the line between primitives
and concrete implementations.

Forth, we have abstractions in Primus Lisp, the notion that is not
representable in the Core Theory (which is not a problem, but raises
a question - shouldn't we consider extending the core theory with
abs/app abstractions)

Finally, we have methods in Primus Lisp, which originally were totally
Primusish stuff, as they were handlers of Primus Observations. Since
BAP Lisp shouldn't depend on Primus (it should be vice verse), we
shall redefine methods somehow. (We don't want to drop them, as it
will render BAP Lisp unusable for Primus - Primus needs methods).
The initial idea, so far, is the methods denote semantics of
effects. Since the notion of the effect is still kind of vague in the
Core Theory, we're open here for interpetations.

If this work will succeed, we will not only get the Primus
interpretation of the core theory, and concrete and parseable BAP
semantics, but we will also be able to write lifters directly in BAP
Lisp.