starts the Primus 2.0 branch
Primus 2.0 will work with the Knowledge and Core Theory Abstractions.

The plan is to factor out the core part of Primus and make it
independent on a particular representation. The BIL dependent
part will go into a separate library primus-bap, while the primus-lisp
will be factored out to bap-lisp. The two former libraries will depend
on primus and will provide the interpreter components. The Primus BIL
Interpreter could be used for backward compatibility, and we're even
going to provide a compatibility library with the former interface.

The core part of Primus will no longer depend on the `project` or
`arch` types, as well as BIL variables or types. Instead we will
use the knowledge library to store the project information, and core
theory for variables and types. The only type from Bap.Std that we're
still using is `word`, which we will soon refactor from bap-std into
a separate library.

During the refactoring, we are going to drop some functionality that
didn't prove useful during the lifetime of Primus 2.0. For example,
we're dropping away generators and various random number generators,
instead employing a much easier to use and support abstractions.

We're also considering to generalizing Primus components and removing
the higher order polymorphism (i.e., defunctorinzing Primus, so that
it no longer will be a transformer). However, this change will involve
a lot of changes in the downstream, so this is still under
consideration.