renames bitvector to bitvec and adds auxiliary libraries
Since we're going to use Bitvec as our next new Bitvector
implementation and since we don't want to alias it to Word or
Addr anymore (because it will introduce confusion with Bap.Std.Word
and Bap.Std.Addr modules as well as significandly slow down
compilation), we need a shorter name. However, Bitv is already used
in the opam universe, and Bv is probably too short. So the middle
ground was chosen, which actually matches with the SmtLib name for
bitvectors.

We also added two sublibraries along with submodules which enable
support for bitvectors in the Janestreet suite of libraries.
- Bitvec_order provides support for the interfaces that depend on
  ordering, like maps, sets, comparables, etc
- Bitvec_binprot provides support bin_io

Together they provide functions that make Bitvec.t a first-class
citizen of JS Core libraries. We're moving away from the old JS style
of providing full comparable or Regular interface for each module,
since it bloats the binaries and, more importantly, significandly
slows down the compilation. We are now relying on small interfaces
such as compartors, so now instead of `Bitvec.Set.t` we will
use `(Bitvec.t, bitvec_order) Set.t` and instead of `Bitvec.Set.Empty`
we will use `Set.empty bitvec_order`

See the module documentation for more information.