switches to an unboxed modulus representation
Since many use cases of the Bitvec interface will use the generic
operations with modulus created on the fly from the integer
represented width, having it boxed will have sever performance
ramifications.

So, instead of representing modulus as a pair of width and the modulus
(i.e., 2^w-1 and w) we have to decide on one or another.

We decided to use 2^w-1 as the modulus representation as this is the
value that is used in normalization, so computing it every time will
add two extra operations which are not [@@noalloc]ed.

The ramification of this choice is that we still need to go back from
2^w-1 to w. The direct solution is to count bits, which is rather
efficient already. For the [msb] operation we don't really need to
know the width to test whether the msb is set, since

     msb x m ::= (m = m lsr 1 lor x) mod m

(though I'm still not sure whichever will be more efficient, the
[test_bit (num_bits m)] or the implementation above, but both are
rather efficient)

We also reimplemented the shift operations. They no longer depend
on the signed extract and are implemented purely relying on bitwise
operations. We use `num_bits` to get the width there.