adds flat domains and made optional flat
A flat domain is stratifies the set into two classes, empty or
something. Empty is the bottom element. All elements that are
not empty are non-comparable.

Basically any set with an empty element satisfies the domain
structure, moreover this is how most of our domains are defined.

Even more, when we want to derive domain from the optional
representation we actually want the flat domain. For example,
the `addr option` representation actually means that there is either
some knowledge about the address of an object, or there is no such
knowlede. It doesn't mean that `Some 0xdead` has more knowledge than
`Some 0xbeef`.