simplifies the class type
Removes the hierarchies, we don't really need them. And when necessary
they could be organized for a concrete family of classes using the
same approach.

It is now possible to declare abstract classes and then refine them to
concrete, or declare a concrete class from the start.

Also, makes the class type constructor covariant, so that a class
could be defined as a polymoprhic instance.