revises the sort system
Implemented a simple algebra of sorts, which consists of sort atoms
and applications of sorts (atoms could be either symbols or numbers).

The algebra is typed, but atoms are universally quantified. A Magic
module is provided which enables breaching of the type system, so
that a sort type could be restored after serialization. This is now
safer, when it was before, as the phantom types is now reconstructred
from the sort definition, rather then set arbitrary at the place of
sort definition.