refines IR jmp representation
Dropes the roles along with multiple links as the were increasing the
impendance mismatch between the representation and its reification
into BIL. The old BIL jmp could have up to two destinations, not N
destinations. Since the new representation was accepting many
destinations it was tempting to use the same jmp term to represent a
control flow which should be represented with several jumps.

In the new represenation we have up to two destinations, which we
denote with `dst` and `alt` detinations. Both are optional, therefore
we now can have 0, 1, or 2 destinations,

```
+-----+-----+-----------+
| dst | alt | semantics |
+-----+-----+-----------+
|  -  |  -  | undefined |
|  x  |  -  | goto x    |
|  -  |  y  | call y    |
|  x  |  y  | call y x  |
+-----+-----+-----------+
```

where `call x` denotes `call x with noreturn` and
`call y x` denotes `call y return x`.

Here we loosing the `ret x` which is denoted with `alt x`
and then reified back to `call x` instead of `ret x`. In order to
recover the call correctly we need to know whether `x` denotes a
function entry or it is in the middle of the function, which is, in
general, an undecidable property. Therefore, we extend the notion of
call, to an interprocedural control flow transfer.

The value of the `dst` and `alt` fields is an abstract variant with
two options, `Indirect vec` parametrized with the vector expression,
denoting a set of possible destinations, and `Resolved tid`,
parametrised with the destination label.