fixes call wiring
as a separate pass (not a real pass though) we rewire all
jmps to entry blocks as calls to corresponding subroutines.
Note, this will also affect recursive subroutines - all jmps
to the self entry block will be treated as interprocedural
jumps. As a nice side effect, this pass will enforce, that the entry
block has not incomming edges.