hardens bap_bitvector implementation
A few bugs were due to bap_bitvector internal representation was
breaching the abstraction of the packed bitvector. The new
implementation carefully establishes boundaries, so that we can't
accidentally treat a packed representation as a bitvector.

Also the textual representation of the Bap_bitvector is restored to
match the previous representation, so that we can pass the tests.