refines reconstructor, symtab, and IR lifter to fit new semantics
Symtab Changes
--------------

First of all, I've refined the symtab interface to make it clearer.
Now we have the shadow callgraph in the symtab, that stores each call
edge from the callsite, denoted with the basic block, to a callee,
which could be a real function also present in the symtab, or an
external, which represented just by its name. We distinguish between
two kinds of calls - explicit and implicit. An explicit call is
produced by a control flow instruction, e.g., jump, branch, or call,
while an implicit call is produced via a fallthrough of the control
flow from one subroutine to another (yes, this can happend, and we
support this since 1.6). Therefore, it is possible that one callsite
has two call edges - one explicit call and another implicit, e.g.,

```
__printf_chk:
    call __stack_check
printf:
    ...
```

Which we should lift as two calls,

```
sub __printf_chk:
    call __stack_check with return next
next:
    call printf
printf:
    ...
```

In general, each block can have 0,1,2 call edges. This is modeled in
the API with two functions, `explicit_callee` and `implicit_callee`,
both returning an optional type. Call edges are inserted into the
callgraph using one `Symtab.insert_call` function, that takes an
optional argument to distinguish between two kinds of edges.

This commit also fixes the remove function, which wasn't removing
function calls, when one was removed.

Reconstructor Changes
---------------------

- Updates to the new symtab interface
- Fixes an externa call classification. It was previously inserted as
  a fall edge, while it is actually an opposite, it is always an
  explicit call.

IR Lifter Changes
-----------------

The IR lifter implementation no longer depends on BIL and is purely
implemented using the new interfaces. It is basically rewritten nearly
from scratch. It is still not passing tests (we still do not insert
synthetic subroutines, they are not needed anymore, but for the sake
of backward compatibility, we will probably introduce them back, maybe
as a separate pass). The new implementation precisely handels
different variants of calls, trying to keep the CFG representation as
tight as possible (i.e., do not introduce unnecessary blocks). It is
also, more conservative wrt to inter/intra classification of
jumps. When the whole program is lifted, we ensure that all goto's
will never leave the boundaries of their subroutines. We turn all
jumps for which we can't provide such guantees to calls. Basically, it
means, that in the output IR it is now impossible to find an
unresolved goto (all gotos now have a form of `goto %....`).
The new lifter is thoroughly commented, so please continue reading
there.