WIP: a complete rew of the recursive descent disassembler
The new implementation is supporting delay slots (and out-of-order
execution in general), more conservative (it disasembles more than a
linear sweep disassembler), uses much less space, and is much easier
to understand and to maintain, since it is only 200 lines of code.

However, there are still some bugs, so it is a work in progress.