changes the scanning mode
now, we only scan as much as needed, no more no less. We no longer
add roots on a failed instruction, but instead, if at the end of
a scan we didn't find all code, we restart we the first unsatisfied
instruction as the root, and so on, until we classify everything as
code or data.