a memory footprint optimization
By using scoped objects for represeting links we can save enomorous
number of bytes. In the previous approach, we were creating a code
object for each by in the input memory and then fill its slots. We
even created a unique symbol for each instruction. Thus, just the
inital markup of 100Mbytes binary takes 50 Gbytes of RAM.

With the new approach it is not only much more memory
efficient (constant space), but also is 4 times faster (probably due
to the less time spent in GC)