adds an optional join method to the domain
This is a big change to the semantics of the knowledge system. Since
now when a piece of knowledge is provided it is joined with the
existing knowledge (if any). E.g., `provide s x p` must be consistent
with the existing `collect s x`, it's now impossible just to override
existing knowledge with something different (well modulo that we can't
really control the consistency of the order and join functions that
are provided by the clients of the library).

When a domain is defined it is possible now to pass a join function
that will provide a least upper bound of two values, or indicate a
conflict. Therefore, now we have a complete lattice stucture for all
properties, with conflict being the top value.

The join functon, if not provided, is inferred from the order
function.

Given the new semantics we now can reach a fixed point in less
iterations and reduce the burden of merging from the knowledge
providers, since now they can rely on a default merge function.

Note, it is still possible to define a recursive property with
arbitrary computation in it, and the system will compute the fixed
point of this computation as previously. It is just possible now to
automatically rely on the default join operator, so, to some extent,
this could be seen as an optimization.

A good example, where we really benefit from the new semantics, is
when we have several independent knowledge providers, like lifters
or jump destination providers. They all are not really interested in
what others are providing, as they just adding their information
(in a different slot, in case of lifter, or by Set.union in case of
dests).