moving forward towards well-defined fixed point semantics
two new updates

1) promises are run only once
2) provide is joining and is forced to preserve monotonicity

Having promises to be computed every time a property is accessed is
stressing the whole system. Ideally, we shall rerun a promise only
if there is some new information, but we don't want to maintain the
whole dependency graph for each promise. Alternatively, a promise
itself may look into the properties and decide whether it has more
information or not. But this will essentially make all promises self
recursive, as the have to first look into the current value, and then
produce better if neccesary. Therefore, we will now evaluate each
property (slot+obj combination) only once (computing the fixed point
as it was before), when the slot of an object is not set. Afterwards,
after the initial value is computed the slot is read and set. Another
problem will re-evaluating promises is that they could every time
generate non-comparable results.

Provide was made monotonic probably in the previous commit. We need
this to enable information joining without introducing expicit
cycles. This change enforces independence on the order of execution,
and ensures fixed point termination. Although it sounds nice, it comes
with a price. Every time we have conflicting opinions (like different
names for the same function) we can't proceed without actually giving
precedence to one or another opinion. Sometimes, consensus just could
not be found, so we have to either create a more complicated lattice
to hold our information, e.g., using sets for all possible names and
aliases of a symbol/use lattice where each name is predicated with
the namer, or, somehow, prune the conflicting information
(thus bringing back to life the problem with the order).

An alternative solutions currently on the table

1) think about the knowledge ordering, like can we treat two different
names, e.g., `.init` and `init_proc` as equivalent from the
perspective of the informational contents. E.g., neither is bearing
more or less information than another. So are they NC or EQ?

2) rewoke the monotonicity requirement, but instead implement joining
only for promises (and probably extend the interface, so that we have
two functions, one that preserves monotonicity and another that
overrides)

3) explore possible data representations. This is the most valid and
theoretically sound approach, but it comes with a huge implementation
and reasoning burden. Like indeed, we have to admit that the same
entity might have different names and in differnt circumstances
different names are applicable. Therefore, a name, instead of being
just a string shall become a (discriminated) union of different
names. E.g., (name,name-provider). We can even enable ranking for
providers, to make it easier to compute a principal name, etc.