renames Bap_state to Bap_toplevel
also Toplevel module was rewritten in the previous commit, we now have
a more top-level like interface. And, we are now having type `'a
Toplevel.t` which denotes the top-level computation monad, which is
declared as `'a Toplevel.t = 'a` in Bap.Std but will be turned back
into a monad outside of Bap.Std.