makes Typeid in Bap.Std.Value abstract, fixes pretty printing
this is the first step, before we will merge Bap.Std.Value into the
new Knowledge framework.

Basically, two options are researched, adding one slot that will
contain the dictionary of all values. Or, turning each tag into
a slot. The latter is more natural, but is harder to implement.