adds x86 SP handling for external calls
So far the stack pointer register wasn't properly handled by
primus in case of external calls.

The reason is in the asymmetry of the function calling in x86 ABI:
the caller side allocates the stack slot for the returning address
and callee sise is responsible for clearing it. But in case of
external calls, dependless if a lisp stub exists or call remains
unresolved, there wasn't any stack restoring. As a result
stack pointer didn't point to a proper place.

This PR solve this problem by incrementing stack pointer
after lisp call returns. It works for unresolved calls as well.