fixes insn delay computation
Instead of providing one promise to compute the semantics of
the instruction we were providing 37 promises (by the number of
branching instructions we support), each promise was asking for
the semantics (for the name), which was triggering the fixed point
computation, therefore we ended up with ~80 computations per
instruction.

Now, we register only one promise. We also request semantics only
after we match the architecture. (In fact, we do not need the
semantics at all, as we could get the name from the program object
itself, though to make mips less dependent on Bap.Std, and for the
sake of demonstration of the Knowledge capabilities, we decided to
keep it).

Note also, that it is fine to request semantics before the arch, it
will end up in a very slight inefficiency, as the delay provider will
be called twice (in case if arch matched, otherwise it will return an
empty instruction and won't be restartred), to make sure that the
fixed point is reached.