revises the class/sort system
it was observed that our slots weren't polymorphic in classes,
therefore we were upcasting our values to base classes, which
wasn't in fact a no-op, since we had to change the class annotation
stored in the value. Therefore, an extra allocation for each Value.get
information.

This forced us to revise the class/sort system. Now, it is not only a
little bit more efficient, but is much more straightforward.

First of all, class is defined as a set of sorts. In terms of OCaml
class is indexed by the sort type variable, so that we can create
multiple instances of the same class parametrized by different sorts.

Slots no longer depend on the sort index, but only on the class index,
therefore all instances of a class, no matter the sort, could be
accessed without any manipulations.

On the Core Theory level we rectify the naming issues and now we have
two sorts: Theory.Value.Sort and Theoyr.Effect.Sort (previously we had
an assymetric sort vs effect, where effect was a sort of effect -
pretty confusing).

We also now have a Theory.Value module, which is symmetric to the
Theory.Effect module, as well as Theory.value and Theory.effect types.
This makes writing types much easier.

Finally, the sort system is also revised, since we now have a notion
of sort on the Knowledge level our Theory sorts fit nicely into
it. They are no longer classes, but, as expected, sorts.

Finally, since we have an ideal empty value for the expression
domain (since the introduction of the Unk type to BIL), we no longer
need to make expression domain optional. This also saves a few
allocations.

As a grain of salt, all those optimizations didn't win a lot, about
10% of performance (and 80% improvemnt in the memory footprint).