yet another pass on dictionary optimization
while our dictionaries are twice as fast as Core's dictionaries for
small inputs, we loose the leadership drastically on larger inputs.

For inputs larger than 1000 we are ten times slowe than our
counterpart :(

A couple of optimizations improved performance, but the overall
picture is still the same.

Two things were optimized - key uid is precomputed and stored in the
key. We use a simple integer instead of relying on the extension
constructor id, since the latter turned out to be rather
expensive (lots of guards).

Second, when a value is inserted we consider cases when it should be
inserted into the head, as previously we were diving down to the leaf
level. And although it was expected that this optimization should fix
the performance gap for large trees it, unfortunately, didn't.

The next suspect is the balancing procedure. An obvious (but tedious)
optimization is instead of creating a disbalanced tree and then
balancing, do this in one step.

Also, it is interesting that core's balance function is not recursive,
which guarantees that there is no more than one balancing per
insertion. Can we guarantee the same?