removes caching from the BIL lifter
Caching on the semantics level breaks the overall semantics of the
program, because semantically equal instructions should still have
different IR (in terms of the labels at least).

It might be indeed an error in the design of the IR lifter, which
probably, should be considered the semantics, but rather a container
for semantics. However, it is not quite obvious how we can separate
them, given that we have to use labels in control flow terms.