switches to custom type keys
the type key from Core (provided by the Type_equal.Id) are relying
on the extension_constructor which is non robust (too many checks)
and is rather slow (again too many checks).

In our representation, we just use our own `int`-typed counter to
check the uniqueness of keys, if needed. In any case OCaml itself
is robust enough to check the keys equality, so we usually don't
need to do its work.