fixes opinions and ignores empty knowledge
So far, we were storing empty knowledge if it was provided directly or
via a promise or suggestion. Besides an increased memory footprint, it
wasn't a big problem, except with the opinions data structure. The
latter weren't doing anything special to handle empty values,
therefore the empty values were treated as normals and were
participating in elections. As a result, if majority of the providers
(or the authorative provider) didn't have the information (provided
the empty knowledge), it, the empty knowledge, was selected.

This commit fixes this issue, in addition we decided to ignore all
attempts to store an empty value in the knowledge base. It is a mere
optimization (for everything except opinions).