revamps the objdump plugin
First of all it fixes a bug when an entry with the name `FUN@plt-OFF`
was considered as a function with name FUN, e.g., both `memove@plt`
and `memmove@plt-0x10` were considered as the `memmove` function which
let to a lot of problems.

Note, this is a normal behavior of objdump, which if the name is not known
just finds the closest known location and adds an offset to it.

Also, since I was here, I gave a little bit of love to the plugin,
notably

- performance optimizations:
  1) process the input on the fly (do not create the list of input,
     lines this is slow and will OOM the plugin for rather large
     input files)
  2) compile the regular expression once (not for every line)
  3) simplify the regexp and get rid of all string ops
     (we can perfectly express what we need in regexp, there is no
      need to apply extra checks or string operations)
  4) get rid of extra allocations (do not create substrings, pairs,
     or lists)
  5) use Z.t instead of words to represent addresses. Takes less
     space, faster, and easier to parse (no need to substring).

- adds an optional demangler
  since objdump can do demangling for us, a new --obdump-demangler
  option controls which demangler is used. Set it to `disabled` to
  disable demangling.