adds support for kernel-style relocation back again
The idea is to fix the fixups as soon as possible to prevent the
problem from spreading to far away. Previously we had to touch a
lot of code to fix the havoc caused by unfixed jumps.

Therefore, the new solution is to fix the BIL code based on the
available relocation information, and everything else will work
correctly automatically.

The main problem here is the representation of the external call,
which we can't represent as a jump, since the latter requires a
concrete address. A simple (but not the cleanesst solution) would
be to use the special statment with a fixup string, which is later
resolved to a correctly named function. Since the solution is totally
confined within the BIL plugin it is rather safe and robust to
changes.

We also added back the synthetic procedures and will keep them until
the abi/api pass is not rewritten in the new framework.