tracks call targets as subroutines
We shouldn't forget to mark targets of known calls as subroutines, so
that our callgraph reconstructor can correctly partition the whole
program CFG.

We are taking an extra care not to mark as subroutines targets of
calls that do not belong to a valid chain (i.e., which will be erased
from a CFG).

The disasm test is also updated to reflect the new changes:

1) The knowledge runtime doesn't allow changing the architecture
of an address, so we had to reset the knowledge.

2) The new CFG reconstruction algorithm is more conservative, so it
now reconstructs an unreachable block, so we had to update the
`call1_3ret` test, in which now our algorithm (correctly) produces a
CFG with 4 blocks.