fixes Primus generators
Due to a design error, Primus self-seeding generators never worked as
expected, as they were recreating themselves on each access. The
problem is more a design error, rather than implementation, therefore
the fix is not very beatiful. It is really hard to implement correctly
the `create init` function which has `(int -> t) -> t` type. The proposed
implementation allows only such init that forall `x, y : int`, `init
x` and `init y` generate generators of the same type.

Probably, there is a better solution (by making the Generator type a
Primus component and turning them into higher-order functions), but
since we dropped generators in Primus 2.0 (in favor of a much easier to
use mechanism), I don't think that it is worthwhile to spend more time
on the current implementation.