fixes taken pathes in the primus promiscuous mode
We had a bug so far when pathes that were taken in
the primus promiscuous mode sometimes were far away
from the real ones.
That was possible because a forked continuation could
return and an execution would continue from a forked
point.

This PR fixes this bug and a forked continuation
never returns.