implements the fair control flow in Primus
As of Primus 1.x and before this commit, the interprocedural control
flow wasn't following the return stack, but instead ignored the return
statements and was always following the control flow graph.

This had several ramifications
1) it was fragile as not all jumps could be properly classified as
returns statically
2) it didn't work with tail calls
3) it wasn't fair, because we were misinterpreting malicious program
as benign.

The new implementation is following the true control flow, i.e., we
are not ignoring any jumps and call the computed destinations. In case
when the dynamic control flow diverges from the static control flow,
we are raise the Cfi_violation exception, which is properly trapped in
the same manner as Division_by_zero and Segmentation_fault.

Implementation
--------------

We use a shadow stack approach, and each interprocedural call which we
expect to return pushes the return value (which we call a prompt) to
that stack. The the contination linked to the call destination is executed
bounded by the pushed prompt. It will stop when it reached the pushed
prompt (i.e., when the control is literally returned to us) or when
the continuation returns normally (e.g., when a Primus Lisp or
primitive is called). In either case, we pop a prompt (if it was
pushed), and compare the prompt value with the expected and raise (a
trappable) exception if they diverge.