fixes insertion of calls for fallthroughs in the IR lifter
Context
=======

Since a subroutine is a set of blocks dominated by its entry point it
is possible to encounter blocks that fall to other subroutines. The
block terminator could a simple move instruction, a (conditional)
jump, or a call to another subroutine.

Let me sprinkle the description with a few example, to make it more
concrete.

Example 1.

Original code:

```
42: mov rax, rbx
44: mov rdi, rsi
```

Suppose that byteweight false positive marked address `44` as a
function start, so we need to lift this code as:

```
rbx := rax
call @sub_44 with noreturn

sub_44:
rsi := rdi
```

This was a simple case, which was working before the PR. Here is a
more convoluted example:

Example 2.
Original code:
```
42: mov rax, rbx
44  call printf
4a: mov rdi, rsi
```

and assume that byteweight introduced a function start at `4a`, so
we now have to lift it as

```
rbx := rax
call @printf return @next

next:
call sub_4a

sub_4a:
rsi := rdi
```

This particular case is handled by this PR. There are also a few cases
in between, when we have a jump or a conditional jump.

To be fair, byteweight is not the only reason why we can have
subroutines split into several parts. The speculative disassembler
will identify piecies of code that are otherwise not reachable (like
switch table destinations, or just pieces of codifiable data inside
functions), therefore this is a quite often case in real binaries, not
a hypothetical situation.

Historical Reference
====================
In BAP 1.6 we were trying to handle all three cases,
but for some reason this erode during the migration, therefore this PR
restores this functionality. Before BAP 2.0 we were using a much less
speculative diassembler, so the only source of such situations was
byteweight. Nowdays, it is more like a norm.

Implementation
==============

We need to distinguish between an explicit call which is comming from
a control flow instruction and an implicit call that comes from a
fallthrough. The last can happen from an explicit call as shown
above. To identify whether we have an explicit call we look into
instruction predicates and also search for interprocedural edges in
our calls graph. If either exists, then the insturction is an explicit
call and we need to turn it into the `call` term. We are also looking
for an implicit interprocedural edge, and if such exists, then we also
need to turn an instruction into the `call` term. In the corner case,
we can have a block that is both an explicit call and implicit call,
and since there is no term in IR which is capable of doing this at the
same time (e.g., `call @malloc return @sub_next`) we have to add an
extra indirection in the form of a block that will do the `sub_next`
call.