relaxes CFI check in the Primus Interpreter
Do not raise a CFI viloation, when the continuation finishes but we
didn't issue any prompts. It usually corresponds just to an execution
of an arbitrary function from a top-level.

This also opens an avenue for making our CFI violation more
informational, since the only left case will have concrete instances
in a form of two destinations - expected vs taken.