parametrize Primus Lisp malloc with initialization strategy
So far our simple-memory-allocator was always filling the allocated
memory with zeros. The proposed implementation adds two parameters
the first one, `*malloc-initialize-memory*` controls whether memory is
initialized at all, and the second one, `*malloc-initial-value*` is
the value that is used to fill in the allocated memory, when the first
parameter is set to `true`.

By default, we set `*malloc-initialize-memory*` to false, so that PR
changes the behavior of Primus. Now the heap memory is randomized, so
if it breaks your analysis the set `*malloc-initialize-memory*` to `true`.