restores the old behavior of the Word.to_int* wrt to signed values
Previously, when a signed bitvector was converted to an OCaml integer
it was sign-extended. This behavior is now restored, e.g.,
```
 # let n_two = Word.of_int (-2) ~width:3;;
 val n_two : Bap.Std.Word.t = 6
 # n_two |> Word.signed |> Word.to_int_exn;;
 - : int = -2
 # n_two |> Word.to_int_exn;;
 - : int = 6
```

Note, also that the new implementation is still not exactly the same
as the old and subsumes the old, for example, in BAP 1.x the following
will raise an error:

```
Word.to_int64_exn (Word.of_int64 0x9a693d39c794d099L);;
```

and therefore the old code was using an explicit `Word.signed` to
trigger the signed extraction, which was probably wrong. In BAP 2.x
no matter whether the value is signed or not, if it fits the target
representation it will be converted, e.g.,
```
 # 0x9a693d39c794d099L;;
 - : int64 = -7320252400943181671
 # Word.of_int64 0x9a693d39c794d099L;;
 - : Bap.Std.Word.t = 0x9A693D39C794D099
 # Word.to_int64_exn (Word.of_int64 0x9a693d39c794d099L);;
 - : int64 = -7320252400943181671
 # Word.to_int64_exn (Word.signed (Word.of_int64 0x9a693d39c794d099L));;
 - : int64 = -7320252400943181671
```